분리된 파일 구조로 변경
