+++ index.html
@@ -13,6 +13,6 @@
 	</noscript>
   <div id="root">
   </div>
-  <script src="app.ts" type="module"></script>
+  <script src="src/app.ts" type="module"></script>
 </body>
 </html>
