+++ src/app.ts
@@ -1,0 +1,27 @@
+import Router from './core/router';
+import { NewsFeedView, NewsDetailView } from './page';
+import { Store } from './types';
+
+const store: Store = {
+  currentPage: 1,
+  feeds: [],
+};
+
+declare global {
+  interface Window {
+    store: Store;
+  }
+}
+
+window.store = store;
+
+const router: Router = new Router();
+const newsFeedView = new NewsFeedView('root');
+const newsDetailView = new NewsDetailView('root');
+
+router.setDefaultPage(newsFeedView);
+
+router.addRoutePath('/page/', newsFeedView, /page\/(\d+)/);
+router.addRoutePath('/show/', newsDetailView, /show\/(\d+)/);
+
+router.go();
