+++ src/page/index.ts
@@ -1,0 +1,2 @@
+export { default as NewsDetailView }  from './news-detail-view';
+export { default as NewsFeedView }  from './news-feed-view';
